Change the interface to EventActions such that external clients can interact with these easily
Notice, this could pose problems if multiple threads are running at the same time. However I'm assuming that
(1) we are only running a single WebKit/WebPage instance at any time and (2) we do not support WebWorkers.

We can change this back when we figure out a good way to schedule and relate multiple related threads.