Use replay identifiers to store the internal list of active event listeners.
Before we used void* pointing to the node, but in some rare cases the node being pointet to were overwritten or somehow moved. This without any call to the destructor (thus we still thought that the node existed).

This avoids this problem by storing a string representation of the node. If the node is gone when we want to trigger it then it will result in a failure to find it, and not a crash.

This case can be observed in https://maps.google.com, specifically for this node.

Running 1921 : 2-AUTO(mousedown, @ ID=inlineTileContainer)
Warning: Node ( @ ID=inlineTileContainer) not found...